Moves SetTable to faust.tables.sets module
