+++ docs/reference/faust.tables.sets.rst
@@ -1,0 +1,11 @@
+=====================================================
+ ``faust.tables.sets``
+=====================================================
+
+.. contents::
+    :local:
+.. currentmodule:: faust.tables.sets
+
+.. automodule:: faust.tables.sets
+    :members:
+    :undoc-members:
